@@ -13,9 +13,9 @@
   "author": "Original Webpack Version by David Ra Youssef (davidradeveloper@gmail.com)",
   "license": "MIT",
   "devDependencies": {
-    "eslint": "^8.4.1",
-    "eslint-plugin-react": "^7.27.1",
-    "prettier": "^2.5.1"
+    "eslint": "^8.13.0",
+    "eslint-plugin-react": "^7.29.4",
+    "prettier": "^2.6.2"
   },
   "keywords": [
     "react",
@@ -25,11 +25,11 @@
   ],
   "dependencies": {
     "react-countdown-clock": "^2.9.0",
-    "@vitejs/plugin-legacy": "^1.6.4",
-    "@vitejs/plugin-react": "^1.1.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "vite": "^2.7.1"
+    "@vitejs/plugin-legacy": "^1.8.0",
+    "@vitejs/plugin-react": "^1.3.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "vite": "^2.9.1"
   }
 }
